Construct the projection pursuit regression tree

Find regression tree structure using various projection pursuit indices in each split.

Usage

PPTreereg(formula,data,DEPTH=NULL,Rr=1,PPmethod="LDA",
                 weight=TRUE,lambda=0.1,r=1,TOL.CV=0.1,selP=NULL,
                 energy=0,maxiter=500,
                 standardized=TRUE,even=TRUE,space=0,
                 maxFinalNode=20,maxNodeN=10,...)

Arguments

formula

an object of class "formula"

data

data frame

DEPTH

depth of the projection pursuit regression tree

Rr

cutoff rule in each node

PPmethod

method for projection pursuit; "LDA", "PDA", "Lr", "GINI", and "ENTROPY".

weight

weight flag in LDA, PDA and Lr index

lambda

lambda in PDA index

r

r in Lr index

TOL.CV

CV limit for the final node

selP

number of variables for the final node in Method 5

energy

energy parameter

maxiter

number of maximum iteration

standardized

standardize each X variable before fitting the tree structure. Default value is TRUE

even

divide evenly at each node. Default value is TRUE

space

space between two groups of dependent variable

maxFinalNode

maximum number of final node

maxNodeN

maximum number of observations in the final node

...

arguments to be passed to methods

Value

Tree.result projection pursuit regression tree result with

PPtreeclass object format

MSE mean squared error of the final tree

mean.G means of the observations in the final node

sd.G standard deviations of the observations in the final node.

coef.G regression coefficients for Method 3, 4 and 5

origY original dependent variable vector

origX.mean mean of original X

origX.sd standard deviation of original X

class.origX.mean means of the each independent variables in the final node

References

...

Examples

data(mtcars)
Tree.result <- PPTreereg(mpg~.,mtcars,DEPTH=2,PPmethod="LDA")
Tree.result
#> ============================================================= 
#> Projection Pursuit Classification Tree result 
#> =============================================================
#> 
#> 1) root
#>    2)  proj1*X <= cut1
#>       4)* proj2*X <= cut2  ->  "1"
#>       5)* proj2*X > cut2  ->  "2"
#>    3)  proj1*X > cut1
#>       6)* proj3*X <= cut3  ->  "3"
#>       7)* proj3*X > cut3  ->  "4"
#> 
#> MSE of various cutoff values 
#> -------------------------------------------------------------
#>      Rule1  Rule2  Rule3  Rule4  Rule5  Rule6  Rule7  Rule8  Rule9
#> MSE 4.3929 4.9249 4.7723 4.3929 4.9249 4.9249 4.9249 4.7723 9.1987