Calculate variable importance

Calculate the importance of variables in the PPTreereg model. For local importance, weighted sum of projection coefficients with the number of data corresponding to each node as the weighted value in each node is used. The global importance is absolute sum of local importance.

Usage

PPimportance(PPTreeregOBJ,...)

Arguments

PPTreeregOBJ

PPTreereg class object - a model to be explained

...

arguments to be passed to methods

Value

An object of the class PPimpobj

Examples

data(dataXY)
Model <- PPTreereg(Y~., data = dataXY, DEPTH = 2)
PPimportance(Model)
#> $imp_node_split
#>   node.id node.n     X1     X2     X3     X4
#> 1       1    100 -0.583  2.414  3.117 -0.346
#> 2       2     50  0.954  3.843 -0.554  0.118
#> 3       3     50 -0.117 -0.462  3.661  1.540
#> 
#> $imp_node_final
#>    FinalNode 1 FinalNode 2 FinalNode 3 FinalNode 4
#> X1       0.053       0.530      -0.262      -0.321
#> X2      -2.168      -0.246       1.322       1.091
#> X3      -1.420      -1.697       0.643       2.474
#> X4       0.143       0.203      -0.558       0.212
#> 
#> $imp_var
#>      X1      X2      X3      X4 
#>  55.916 228.309 261.208  58.741 
#> 
#> attr(,"class")
#> [1] "PPimportance"